+++ tests/testthat/test-rsdqcs.R
@@ -1,13 +1,10 @@
 library(testthat)
 suppressPackageStartupMessages(library(mzQuality))
-# Read the example dataset
-dataFile <- system.file("extdata/data.RDS", package = "mzQuality")
-exp <- readRDS(dataFile)
-
+path <- system.file("extdata", "example.tsv", package = "mzQuality")
+exp <- buildExperiment(readData(path))
 
 test_that("Internal Standards can be replaced", {
     exp2 <- replaceInternalStandards(exp)
-    expect_true(is(exp, "SummarizedExperiment"))
-
+    expect_true(isValidExperiment(exp))
     expect_true(all(rowData(exp2)$suggestedIS == rowData(exp2)$compound_is))
 })