+++ NEWS.md
@@ -2,11 +2,17 @@
 - Moved data files from inst/* to inst/extdata/* .
 - Added description of data files in inst/script/README.md.
 - Changed extensions of .txt files to .tsv to reflect the tab-separated format.
+- Removed .RDS example file in favor of a .tsv file. Consequently, the 
+example documentation and tests have been altered to read the .tsv file.
 - Moved the files in inst/rmd/* to inst/templates/* and renamed them to
   reflect their purpose.
 - Added `mzQuality` vignette, showcasing a typical mzQuality workflow.
 - Removed `doAll` and `aliquots` arguments from `doAnalysis` as they were
-  no longer used
+  no longer used.
+- Renamed `rowIndex` and `colIndex` arguments in `buildExperiment` to 
+`compoundColumn` and `aliquotColumn` to make them more intuitive.
+- fixed a bug where a single file containing multiple batches was recognized 
+as a single batch.
 
 # mzQuality 0.99.1